replace tagged-tokens with addressable
an addressable is a collection of labels, an instruction, a collection arguments or data, and a collection of comments, which is declared as a single block in the assembly source.

point of contention is the comments, should comments before an addressable be included in that addressable, and should comments after an addressable be included?

currently comments after an addressable are included with it until a label is found.